Start of Session Dealing
Modified the footer to use HTML entity for copyright logo, as the UTF-8
character would not properly appear. Added in a simple login check
function, as well as a function that would protect the page from users
who are not logged in. Added the initial login check for the navigation
menu in the header file.